More dockertools cleanup
Move some constants/functions to dockershim and remove unused tests.
