+++ pkg/kubelet/dockershim/naming.go
@@ -23,7 +23,6 @@ import (
 	"strings"
 
 	runtimeapi "k8s.io/kubernetes/pkg/kubelet/api/v1alpha1/runtime"
-	"k8s.io/kubernetes/pkg/kubelet/dockertools"
 	"k8s.io/kubernetes/pkg/kubelet/leaky"
 )
 
@@ -51,9 +50,9 @@ const (
 	// Delimiter used to construct docker container names.
 	nameDelimiter = "_"
 	// DockerImageIDPrefix is the prefix of image id in container status.
-	DockerImageIDPrefix = dockertools.DockerPrefix
+	DockerImageIDPrefix = "docker://"
 	// DockerPullableImageIDPrefix is the prefix of pullable image id in container status.
-	DockerPullableImageIDPrefix = dockertools.DockerPullablePrefix
+	DockerPullableImageIDPrefix = "docker-pullable://"
 )
 
 func makeSandboxName(s *runtimeapi.PodSandboxConfig) string {
